+++ apps/agentic-server/src/mastra/tools/asset/constants.ts
@@ -1,0 +1,13 @@
+// Global unified network identifiers used across all asset tools
+export const UNIFIED_NETWORKS = [
+  'ethereum',
+  'optimism',
+  'arbitrum',
+  'polygon',
+  'avalanche',
+  'bsc',
+  'base',
+  'gnosis',
+] as const
+
+export type UnifiedNetwork = (typeof UNIFIED_NETWORKS)[number]
